--- src/component/Gallery/GalleryLayout.css
@@ -1,24 +1,0 @@
-.photos-container {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 16px;
-  padding: 20px;
-}
-
-@media (min-width: 769px) and (max-width: 900px) {
-  .photos-container {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-@media (min-width: 500px) and (max-width: 768px) {
-  .photos-container {
-    grid-template-columns: repeat(2, 1fr) !important;
-  }
-}
-
-@media (max-width: 499px) {
-  .photos-container {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
