+++ src/component/GalleryList/PhotoDetails.css
