+++ src/component/GalleryList/PhotoDetails.jsx
