+++ src/component/GalleryList/toggleFavourite.js
