+++ src/pages/Favourite/index.jsx
@@ -1,0 +1,37 @@
+import React, { useState, useEffect } from "react";
+import GalleryLayout from "../../component/GalleryList";
+import toggleFavourite from "../../component/GalleryList/toggleFavourite";
+
+const FavouritePhotos = () => {
+  const [favouritePhotos, setFavouritePhotos] = useState([]);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    try {
+      const storedFavourites =
+        JSON.parse(localStorage.getItem("favourites")) || [];
+      setFavouritePhotos(storedFavourites);
+    } catch (error) {
+      console.error("Error fetching favourite photos:", error);
+      setError("Error fetching data!");
+    }
+  }, []);
+
+  const handleToggleFavourite = (id) => {
+    const updatedFavourites = toggleFavourite(favouritePhotos, id);
+    setFavouritePhotos(updatedFavourites);
+    localStorage.setItem(
+      "favourites",
+      JSON.stringify(updatedFavourites.filter((photo) => photo.favourites))
+    );
+  };
+  return (
+    <GalleryLayout
+      photos={favouritePhotos}
+      error={error}
+      onToggleFavourite={handleToggleFavourite}
+    />
+  );
+};
+
+export default FavouritePhotos;
