--- src/pages/FavouritePage/index.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Favourites from "../../component/Favourites";
-
-function favouritePhotos() {
-  return (
-    <>
-      <Favourites />
-    </>
-  );
-}
-
-export default favouritePhotos;
